Allow safe usage of OSAL api in interrupt context
The OSAL api uses a global lock to perform some operations using mutex,
semaphores, queues and so. This leads to a possible deadlock in the program
due the following situation.

  Let "I" be the timer interrupt (SIG63):

       |[Thread 1]
       |  ...
       |  call OS_TaskGetInfo
       |    ...
       |    lock(global)
       |    ...
  I----+--------------------+
                            |[Interrupt]
                            |  ...
                            |  call OS_TaskGetInfo
                            |    ...
                            |    lock(global)  // DEADLOCK

There are some possible solutions to avoid this problem. One is to never use any
OSAL primitives on the interrupt context, which quite limit the library. Other
would be block all signals during the global lock and let them be pending. This
approach may introduce poor performance. Other would be find out how to remove
these global locks. Another could be use safe locks on interrupt handler.

This patch implements the blocking signals solution because it is the best
trade-off of the current solutions.

There is also a single special case which is not handled by this patch.

Signed-off-by: Vinicius Tinti <viniciustinti@gmail.com>