Fix #648, scrub array references in shared layer
Introduce the OS_object_token_t type which tracks a reference
to an OSAL resource.  This contains all information about the
original reference, including the ID, object type, lock type, and
the table index.

Therefore, since the token type contains all relevant info, it
can be used in all places where a bare index was used.

This also considerably simplifies the code, as some functions
which previously output multiple objects only need to operate on
tokens, and the functions which called these functions only need
to instantiate a token.