Fix #648, rename internal fields for consistency
Use a variable name that matches the type of resource being accessed,
rather than just "local".  In particular this is important for readability
of timecb and timebase code where functions need often need to access both
types of objects at the same time.  This also updates filesys code to match.