Fix #703, implement exclusive/reserved lock
Change the EXCLUSIVE lock type such that it sets the ID in the global
table to RESERVED and unlocks the global before returning to the caller.

This allows the potentially long-running operation to complete and
not block other operations from happening in other tasks.

Use the EXCLUSIVE lock for all create/delete ops as well as for
bind and connect socket ops.

Also implement a new "RESERVED" lock to handle a special case in the
vxworks timebase implementation where the impl layer needs to acquire
a token for an object as it is being created.  This case is special
because it needs to happen during OS_TimeBaseCreate, and cannot be
completed after the fact like normal tasks, because it is a factor
in determining the success/fail status of the overall operation.