Fix #889, report timer_gettime error
In VxWorks the impl calls timer_settime followed by timer_gettime on the
same timer to check if rounding occurred.  If the second call fails this
reports it as an error.

Note unless there is some sort of OS bug this should be impossible to
happen as this code only runs after a successful timer_settime.