Fix #863, check/report fcntl status
The fcntl() function is documented as returning -1 on error, so
test for this condition and report errno if so.  There is no
recourse/handling - this just reports the error.

However - failure to set the O_NONBLOCK flag via this method
will fall back to blocking mode being used (timeouts will not
work as a result).