Fix #844, switch to use CLOCK_REALTIME
The portable clock_gettime implementation had been using CLOCK_MONOTONIC
to support its use as a PSP timebase for some platforms that used it this
way.  However with updates on the PSP side this is not required anymore.

Preference should be to use CLOCK_REALTIME as it better aligns with the
described semantics of the OSAL clock function, and makes for a better
default.  This can still be easily changed back if the user desires.