Fix #939, rework shell test
The shell may add extra output beyond the intended content, but the
shell-test was only looking for exact match.  Extra output may
include whitespace and/or an echo of the command itself.

This reworks the shell test to allow for this extra output, as long
as the intended string appears in the content of the file it will
pass.