Fix #964, UtPrintx function
Fix the UtPrintx() routine such that the loop stops correctly.
Also improves the output to print the address, not just the data.

Note if UtPrintf is used, one sees the file/line of the UtPrintx
function, not the actual test location, so it is better to call
UT_BSP_DoText directly so it omits this extraneous info.