Fix #969, Use UtAssert tool API for count pattern
The stream test was sending an 8-bit count pattern to validate
the connection, but was not using the provided API to do so.

This provides the exact same test but uses the provided API
from UtAssert to implement it, rather than duplicating the logic.

Notably, this also confirms that the tool functions work as
expected, since there were no other OSAL tests using this API.