Fix #640, configuration guide updates
Update the configuration guide for the current build system (CMake)
and the currently provided set of OS/BSP layers.  Many of the
examples in this guide referred to items which are obsolete.

However, this does *NOT* fix references to the applications and
in particular the "tasking example" as this still may be brought back.