Fix #988, do not require nonblock mode
In some versions of VxWorks the fcntl F_GETFL/F_SETFL opcodes
do not appear to be implemented, and thus it is not possible
to set O_NONBLOCK mode.  However, this mode is not necessarily
required, it is more of a backup/failsafe.

The "selectable" flag should not be dependent on whether
O_NONBLOCK flag got set.

This also adjust some timeouts and adds some delays to improve
the reliability of network-api-test on VxWorks.  The timeouts
were only 10ms, and this is much too short as messages are
getting written on a 9600 baud console (avg 1ms/char).  A
single log message can easily take 50-60ms alone.