Fix #1024, check misc API return codes
Fix missing doxygen retvals

Note, other tests were OK, but depend on support being enabled and
supported by RTOS.  Script will report N/A items (due to no impl) as
missing.  Confirmed OS_HeapGetInfo OK on RTEMS.

For test cases where a function is invoked before OS_API_Init, only
check that the result is not success.  This will remain as "undocumented"
behavior - users should not call functions before init, because it is
not valid to do so.  A specific error code is not guaranteed here.