Fix #1082, increase UT symbol dump size limit
For the MCP750, the symbol table fairly large and requires a considerably
larger limit to avoid the OS_ERR_OUTPUT_TOO_LARGE limit.  This also uses
a different name, so that the user can actually see the symbol dump file
if they want (otherwise the next test overwrites the file).