Fix #1088, avoid task delete during UtPrintf
In the osal_core_UT test, the test task was being deleted while it was
in the midst of a UtPrintf call, which left the BSP mutex in a locked
state, causing deadlock.

Using SEM_DELETE_SAFE attribute avoids task deletion for the mutex
holder, and adding a small delay to the test case makes all the
messages appear as expected (accounts for slow serial console on
test platform).