Fix #1124, add generic asserts from CFE coverage testing
Migrates some of the assert extensions that were added for CFE
coverage testing to the UtAssert level, so they can be used
in a wider scope of tests, not just CFE coverage.

The existing macros (e.g. UtAssert_INT32_EQ) are updated to
use the same facility.

All macros are now also a single line and return the boolean
pass/fail status, which was a useful feature of the CFE macros.