Fix #1130, add test case types similar to NA
Add two more test case variants, similar to NA, where failure
of the test case does not translate to failure of the overall
test:

 - UTASSERT_CASETYPE_WARN
 - UTASSERT_CASETYPE_FLOW

The additional cases have different levels of default visibility,
and may be handled differently by the end user.  That is, some
tests may be skipped because they are truly NA (and nothing for
the user to do to change that) and some tests may be skipped
because the system was not set up in a way that allowed them
to be run (and the user must fix that and re-run).  For the
latter case, the WARN type can be used, to more clearly
indicate it is an action item for the user.

Lastly the "FLOW" message is intended to indicate internal decisions
in the test case implementation (not actual test cases).