Fix #1199, correct warnings on gcc11
These new warnings are detected by the new compiler. They are all places
in unit test where an uninitialized value is passed via "const" pointer
into a unit under test.  By definition a "const" pointer is always an
input.

While the warning is pedantically true - should not pass an uninitialized
struct as an input - these were all unit tests where the object value
is a don't-care value, so it does not matter.  But to fix the warning,
simply need to initialize a value before making the call.