Fix #1304, locks for condvar objects on rtems/vxworks
Instantiate the condvar lock on rtems and vxworks even though the
feature is not currently implemented.  This is because OS_MAX_CONDVARS
must still be nonzero and if the "Create" function is called, it still
goes through the motions of allocating an ID.

This also adds a NULL check in the lock implementation, and thus if a
lock is not instantiated in the future this will not dereference a NULL
pointer.