Fix #1318, adjust timing on condvar tests
The order of task wakeup when using OS_CondVarSignal() with multiple
tasks pending may not be guaranteed.  This relaxes the test case so that
a specific order is not assumed/required.  This also adjust certain
sleep times to make sure the test waits long enough for the tasks to do
the expected work.