Fix #1339, add doc-prebuild dependency
When building as part of CFE/CFS, register the osal header list file
(osal-public-api.doxyfile) as a dependency of the doc-prebuild abstract
target.  This way, CFE and any other CFS apps that refer to the OSAL
headers in their documentation can easily ensure that this (and any
other file produced by other modules) exist prior to attempting to run
doxygen.