Fix #1344, UT assert stub return values
Improves default support for return values of sizes other than 32 bits.
Previously this required a custom handler for any value that was not a
32 bit integer.  With this update any integer value can be used, and a
value translation will occur if the size is differnt.  Thus boolean
values, 8,16, and 64-bit ints can be configured through the same basic
API without needing to add a handler.