Fix #1173, separate append on volume_name to system_mountpt
Generating the system_mountpt string via a single call to snprintf
triggered a compiler warning about overlapping memory.  However this
does not seem like a real warning, as the volume_name should always
be null terminated before the overlap would be possible.

By separating this to be a separate append of the volume_name along with
an explicit size check to ensure the buffers indeed do not overlap, this
avoids the warning.