Fix #1365, fixup API headers for C++
Correct a few syntactical items that are valid C99 but not valid for
C++03.  This includes use of the designated initializer, compound
literals, variadic macros, and a trailing comma in enum declarations.

Note this only matters for public API headers, as some users may need to
include these directly from a c++ file.  For normal C source files, these
are all compiled as C99 and thus the syntax is fine.