Fix #1433, use virtual path as name for FS_BASED maps
Use the "virtual_path" instead of "device_name" as the name
record for FS_BASED entries.  This should allow duplicate
detection to work as expected.

The device_name on mappings made via this method will now be
an empty string.  It is not relevant because this type of map
is used for devices that are already mounted.