Fix #1445, add conditional around priority sem test
The multi-thread sem tests (count + bin) have a test that relies on the
priority taking effect, that the highest priority thread will get the
sem first.  However when running as non-root and in permissive mode, the
priority is not applied and thus this becomes a race condtion.

This adds a test and skips the test if priority is not enabled