Fix #1516, Patch virtual mount point checking
Sponsored by 21Software

Reject any paths which might be part of a path traversal attack by
matching directory entries such as `.` and `..` at the end of the given
path. This hardens a vulnerability where an attacker might be able to
write files in unauthorized locations outside of a virtual mount point
directory.